requirements.txt
fastapi
uvicorn
langchain
langchain-community
sentence-transformers
pinecone-client
python-dotenv
pydantic
boto3
